+++ internal/resource/config/passwordcredentialvalidator/password_credential_validator_resource.go
@@ -146,7 +146,7 @@ func (r *passwordCredentialValidatorResource) ValidateConfig(ctx context.Context
 	var isRadiusServerTableFound bool
 	if pluginDescriptorRefId == "org.sourceid.saml20.domain.RadiusUsernamePasswordCredentialValidator" || pluginDescriptorRefId == "org.sourceid.saml20.domain.SimpleUsernamePasswordCredentialValidator" {
 		if configuration["tables"] != nil {
-			tables := configuration["tables"].(types.Set).Elements()
+			tables := configuration["tables"].(types.List).Elements()
 			for _, table := range tables {
 				tableAttrs := table.(types.Object).Attributes()
 				tableName := tableAttrs["name"].(types.String).ValueString()
@@ -156,10 +156,24 @@ func (r *passwordCredentialValidatorResource) ValidateConfig(ctx context.Context
 					for tableRowIndex, row := range tableRow {
 						rowAttrs := row.(types.Object).Attributes()
 						fields := rowAttrs["fields"].(types.Set).Elements()
+						sensitiveFields := rowAttrs["sensitive_fields"].(types.Set).Elements()
 						usernameFound := false
 						passwordFound := false
 						confirmPasswordFound := false
 						for _, field := range fields {
+							fieldRow := field.(types.Object).Attributes()
+							nestedTableFieldName := fieldRow["name"].(types.String).ValueString()
+							if nestedTableFieldName == "Username" {
+								usernameFound = true
+							}
+							if nestedTableFieldName == "Password" {
+								passwordFound = true
+							}
+							if nestedTableFieldName == "Confirm Password" {
+								confirmPasswordFound = true
+							}
+						}
+						for _, field := range sensitiveFields {
 							fieldRow := field.(types.Object).Attributes()
 							nestedTableFieldName := fieldRow["name"].(types.String).ValueString()
 							if nestedTableFieldName == "Username" {
@@ -211,7 +225,13 @@ func (r *passwordCredentialValidatorResource) ValidateConfig(ctx context.Context
 	fieldNameMap := map[string]bool{}
 	if configuration["fields"] != nil {
 		fields := configuration["fields"].(types.Set).Elements()
+		sensitiveFields := configuration["sensitive_fields"].(types.Set).Elements()
 		for _, field := range fields {
+			field := field.(types.Object).Attributes()
+			fieldName := field["name"].(types.String).ValueString()
+			fieldNameMap[fieldName] = true
+		}
+		for _, field := range sensitiveFields {
 			field := field.(types.Object).Attributes()
 			fieldName := field["name"].(types.String).ValueString()
 			fieldNameMap[fieldName] = true
@@ -337,8 +357,7 @@ func (r *passwordCredentialValidatorResource) Create(ctx context.Context, req re
 	}
 
 	// Configuration
-	configuration := client.NewPluginConfigurationWithDefaults()
-	err = json.Unmarshal([]byte(internaljson.FromValue(plan.Configuration, true)), configuration)
+	configuration, err := pluginconfiguration.ClientStruct(plan.Configuration)
 	if err != nil {
 		resp.Diagnostics.AddError(providererror.InternalProviderError, "Failed to build plugin configuration request object: "+err.Error())
 		return
@@ -416,8 +435,7 @@ func (r *passwordCredentialValidatorResource) Update(ctx context.Context, req re
 	}
 
 	// Configuration
-	configuration := client.NewPluginConfiguration()
-	err = json.Unmarshal([]byte(internaljson.FromValue(plan.Configuration, true)), configuration)
+	configuration, err := pluginconfiguration.ClientStruct(plan.Configuration)
 	if err != nil {
 		resp.Diagnostics.AddError(providererror.InternalProviderError, "Failed to build plugin configuration request object: "+err.Error())
 		return